Refactor auth module structure and improve code formatting
- Move AuthService to features/auth/services directory
- Add auth routing configuration with lazy-loaded login/register components
- Create AuthStateService in auth module
- Format function parameters in auth interceptor for better readability
