+++ src/app/features/auth/auth.routes.ts
@@ -1,0 +1,12 @@
+import { Routes } from "@angular/router";
+
+export const routes: Routes = [
+  {
+    path: 'login',
+    loadComponent: () => import('./components/login/login.component').then(c => c.LoginComponent)
+  },
+  {
+    path: 'register',
+    loadComponent: () => import('./components/register/register.component').then(c => c.RegisterComponent)
+  }
+];
