+++ src/app/features/auth/services/auth-state.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class AuthStateService {
+
+  constructor() { }
+}
